+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Breaking changes:
 - Use the same message for success and error HTTP logs.
 - Remove `req`, `res`, `serviceContext`, and `responseTime` from prettified logs when debugging.
 
+Features:
+
+- Accept an optional side effect function with `toNull`.
+
 ## v1.0.0-rc.4 (2025-07-17)
 
 Breaking changes: